feat: merge to main.py
- merge fire functions to main.py
