+++ src/dataset/data_process.py
@@ -239,60 +239,5 @@ def get_dataset(df, folds_index):
         fold_datasets.append((train_dataset, val_dataset))
     return fold_datasets
 
-def run_geoprocess():
-    from dotenv import load_dotenv
-    import pandas as pd
-    import os
-
-    load_dotenv(dotenv_path=os.path.join(project_path(), '.env'))
-    remote_raw_datapath = os.getenv("S3_URL")
-    # 데이터 파이프라인 주기에 맞춰 다운로드 URL 수정
-    remote_raw_datapath = remote_raw_datapath.replace(".csv", f"_{get_current_time(strformat='%y%m%d')}.csv")
-    # download apt raw data
-    apt = pd.read_csv(remote_raw_datapath)
-    # process columns
-    apt = apt_preprocess(apt, only_column=True)
-    apt_unique = get_unique_apt(apt)
-    for _ in range(5):
-        # web crawling을 통해 좌표 검색
-        apt_unique = get_location_dataframe(apt_unique, num_workers=8)
-        if apt_unique[apt_unique['X']==0].shape[0] < 1:
-            break
-    # 좌표 데이터프레임을 S3에 업로드
-    apt_unique = save_location_s3(apt_unique)
-
-def run_preprocess():
-    from dotenv import load_dotenv
-    import pandas as pd
-    import os
-
-    load_dotenv(dotenv_path=os.path.join(project_path(), '.env'))
-    remote_raw_url = os.getenv("S3_URL")
-    # 데이터 파이프라인 주기에 맞춰 다운로드 URL 수정
-    remote_raw_url = remote_raw_url.replace(".csv", f"_{get_current_time(strformat='%y%m%d')}.csv")
-    # read data from S3
-    apt = read_remote_dataset(remote_raw_url)
-    # preprocess 
-    apt = apt_preprocess(apt)
-    # upload processed apt dataframe to S3
-    upload_url = os.getenv("S3_APT_PROCESSED")
-    upload_url = upload_url.replace(".csv", f"_{get_current_time(strformat='%y%m%d')}.csv")
-    print(apt.columns)
-    print(apt.isnull().sum())
-    apt.to_csv(upload_url, index=False)
-
 if __name__ == '__main__':
-    import sys
-    sys.path.append(
-        os.path.dirname(os.path.dirname( # /mlops/
-            os.path.dirname(  # /mlops/src
-                os.path.abspath(__file__)  # /mlops/src/main.py
-            )
-        ))
-    )
-
-    from src.dataset.data_geoprocess import (
-        save_location_s3, get_location_dataframe, get_unique_apt
-    )
-    # run_geoprocess()
-    # run_preprocess()
+    pass
